refactor: ファイルを指定ディレクトリに移動
